Fix implicit narrowing conversion in compound assignment reported by CodeQL
Explicitly cast to the required type, the possible information loss it
fine.

+++ server/src/main/java/io/crate/protocols/postgres/types/NumericType.java
@@ -127,7 +127,7 @@ class NumericType extends PGType<BigDecimal> {
             for (var nextDigitIdx = digitIdx + DEC_DIGITS; digitIdx < nextDigitIdx; digitIdx++) {
                 ndigit *= 10;
                 if (digitIdx >= start && digitIdx < end) {
-                    ndigit += digits[digitIdx] - '0';
+                    ndigit += (short) (digits[digitIdx] - '0');
                 }
             }
             buffer.writeShort(ndigit);
